+++ apps/dialtone-documentation/docs/.vuepress/baseComponents/constants.js
@@ -1,1 +1,1 @@
-export const ICON_KINDS = ['spot'];
+export const ICON_KINDS = ['spot', 'brand-logos'];