+++ packages/dialtone-vue2/.storybook/main.js
@@ -3,7 +3,7 @@ import { mergeConfig } from 'vite';
 /** @type { import('@storybook/vue-vite').StorybookConfig } */
 const config = {
   stories: [
-    '../@(components|directives|recipes)/**/*.stories.@(js|jsx|ts|tsx)',
+    '../@(components|directives|recipes|prototypes)/**/*.stories.@(js|jsx|ts|tsx)',
     '../@(components|directives|docs|functions|recipes)/**/*.mdx',
   ],
   addons: ["@storybook/addon-links", "@storybook/addon-essentials", "@storybook/addon-a11y", 'storybook-dark-mode'],