+++ packages/dialtone-vue2/prototypes/dialpad/dialpad_default.story.vue
@@ -1,0 +1,20 @@
+<template>
+  <dialpad-layout>
+    <template #leftbar>
+      <dialpad-leftbar />
+    </template>
+  </dialpad-layout>
+</template>
+
+<script>
+import DialpadLayout from './dialpad_layout.vue';
+import DialpadLeftbar from './leftbar/dialpad_leftbar.vue';
+
+export default {
+  name: 'DialpadDefault',
+  components: {
+    DialpadLayout,
+    DialpadLeftbar,
+  },
+};
+</script>