+++ packages/dialtone-icons/CHANGELOG.md
@@ -1,3 +1,85 @@
+# [4.24.0](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.23.0...dialtone-icons/v4.24.0) (2024-07-03)
+
+
+### Bug Fixes
+
+* NO-JIRA keywords ([eb8527d](https://github.com/dialpad/dialtone/commit/eb8527d12255707d93425e0719595b029c2fa2f5))
+
+
+### Features
+
+* **Rich Text Editor:** DLT-1835 add undo behaviour ([#390](https://github.com/dialpad/dialtone/issues/390)) ([2d65590](https://github.com/dialpad/dialtone/commit/2d65590147c0a3f24e6294e23e17aff5beab6b27))
+
+# [4.23.0](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.22.0...dialtone-icons/v4.23.0) (2024-07-02)
+
+
+### Features
+
+* **Icon:** DLT-1804 add t-mobile ([#387](https://github.com/dialpad/dialtone/issues/387)) ([4ce82b2](https://github.com/dialpad/dialtone/commit/4ce82b21b2466b2056d802a6b6cf6f936dd09091))
+
+# [4.22.0](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.21.0...dialtone-icons/v4.22.0) (2024-07-01)
+
+
+### Bug Fixes
+
+* **Emoji Picker:** DP-102475 fix keyboard navigation ([#389](https://github.com/dialpad/dialtone/issues/389)) ([6ca1d16](https://github.com/dialpad/dialtone/commit/6ca1d16b934f377414c652d102429c57fb76182b))
+* **Logo:** DLT-1833 corrected antialiased edging  ([#381](https://github.com/dialpad/dialtone/issues/381)) ([67322e6](https://github.com/dialpad/dialtone/commit/67322e69bab916a403387159187cb2a01b15f0ae))
+
+
+### Features
+
+* **Icon:** DLT-1712 add token icon  ([#388](https://github.com/dialpad/dialtone/issues/388)) ([66d2474](https://github.com/dialpad/dialtone/commit/66d2474afab14373d9f75235454995fba929f52f))
+
+# [4.21.0](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.20.2...dialtone-icons/v4.21.0) (2024-06-25)
+
+
+### Features
+
+* **Icon:** DLT-1806 accessibility-mac ([#365](https://github.com/dialpad/dialtone/issues/365)) ([8735843](https://github.com/dialpad/dialtone/commit/8735843735d5fd2b8c39084a7d2c553329428f73))
+
+## [4.20.2](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.20.1...dialtone-icons/v4.20.2) (2024-06-18)
+
+
+### Bug Fixes
+
+* **Icons:** NO-JIRA add webex.svg ([14ede84](https://github.com/dialpad/dialtone/commit/14ede84a4111c8e0d832992c6c389791c12a5ee6))
+* **Icons:** NO-JIRA delete webex.svg ([94162dc](https://github.com/dialpad/dialtone/commit/94162dcf5995b828de394ab6277370d400eb134c))
+
+## [4.20.1](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.20.0...dialtone-icons/v4.20.1) (2024-06-18)
+
+
+### Bug Fixes
+
+* **Icons:** NO-JIRA webex icon ([aecc85b](https://github.com/dialpad/dialtone/commit/aecc85b10a0a1e5743e29ef19f43dc61b5b57d5e))
+
+# [4.20.0](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.19.0...dialtone-icons/v4.20.0) (2024-06-18)
+
+
+### Features
+
+* **Icon:** DLT-1804 add webex ([#367](https://github.com/dialpad/dialtone/issues/367)) ([72cab26](https://github.com/dialpad/dialtone/commit/72cab26d6ff379cd5ca2b6610ace5465b6d7190b))
+
+# [4.19.0](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.18.0...dialtone-icons/v4.19.0) (2024-06-18)
+
+
+### Features
+
+* **Logo:** NO-JIRA gradient update ([#372](https://github.com/dialpad/dialtone/issues/372)) ([559a517](https://github.com/dialpad/dialtone/commit/559a5178131f799065cbdcc5610238cac9d9556e))
+
+# [4.18.0](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.17.1...dialtone-icons/v4.18.0) (2024-06-13)
+
+
+### Features
+
+* **Logo:** DLT-1802 color options ([#359](https://github.com/dialpad/dialtone/issues/359)) ([14958c8](https://github.com/dialpad/dialtone/commit/14958c8629b2d0697abacf3e9bcfe00b6452442c))
+
+## [4.17.1](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.17.0...dialtone-icons/v4.17.1) (2024-06-07)
+
+
+### Bug Fixes
+
+* **Icon:** DLT-1785 dialpad-ai-color-reversed 1px too big ([#349](https://github.com/dialpad/dialtone/issues/349)) ([55a1ca2](https://github.com/dialpad/dialtone/commit/55a1ca28f28b0efc130a3471423a6a8f03d3f9d4))
+
 # [4.17.0](https://github.com/dialpad/dialtone/compare/dialtone-icons/v4.16.3...dialtone-icons/v4.17.0) (2024-06-04)